Refactor logging
This commit refactors logging from a special `LoggingNetwork`
subclass into the `DefaultNetwork` base class.

In addition, this commit fixes `setup_logging` so that calling
it multiple times does not add additional handlers.

The special `LoggingClient` subclass now does nothing more than
call `setup_logging`, since it doesn't need to use a special
network subclass.
