+++ boxsdk/__init__.py
@@ -5,4 +5,5 @@ from __future__ import unicode_literals, absolute_import
 from .auth import JWTAuth, OAuth2
 from .client import *  # pylint:disable=wildcard-import,redefined-builtin
 from .object import *  # pylint:disable=wildcard-import,redefined-builtin
+from .util.log import setup_logging
 from .version import __version__
