+++ boxsdk/client/logging_client.py
@@ -3,16 +3,13 @@
 from __future__ import unicode_literals, absolute_import
 
 from .client import Client
-from ..network.logging_network import LoggingNetwork
+from ..util.log import setup_logging
 
 
 class LoggingClient(Client):
     """
     Box client subclass which logs requests and responses.
     """
-    def __init__(self, oauth=None, network_layer=None, session=None):
-        super(LoggingClient, self).__init__(
-            oauth=oauth,
-            network_layer=network_layer or LoggingNetwork(),
-            session=session,
-        )
+    def __init__(self, *args, **kwargs):
+        setup_logging()
+        super(LoggingClient, self).__init__(*args, **kwargs)
