+++ demo/cooperative_auth.py
@@ -2,16 +2,15 @@
 
 from __future__ import unicode_literals, absolute_import
 
-from logging import Logger
+from logging import getLogger
 from multiprocessing import Manager, Process
 from os import getpid
 
 from boxsdk.auth.cooperatively_managed_oauth2 import CooperativelyManagedOAuth2
-from boxsdk.network.logging_network import LoggingNetwork
 from boxsdk.util.log import setup_logging
 from boxsdk import Client
 
-from auth import authenticate, CLIENT_ID, CLIENT_SECRET
+from .auth import authenticate, CLIENT_ID, CLIENT_SECRET
 
 
 def main():
@@ -44,9 +43,8 @@ def _store_tokens(access_token, refresh_token):
 
 def worker():
     # Set up a logging network, but use the LoggingProxy so we can see which PID is generating messages
-    logger = setup_logging(name='boxsdk.network.{0}'.format(getpid()))
-    logger_proxy = LoggerProxy(logger)
-    logging_network = LoggingNetwork(logger)
+    logger = getLogger('boxsdk.network.{0}'.format(getpid()))
+    setup_logging(name=logger.name)
 
     # Create a coop oauth2 instance.
     # Tokens will be retrieved from and stored to the multiprocessing Namespace.
@@ -58,12 +56,11 @@ def worker():
         client_id=CLIENT_ID,
         client_secret=CLIENT_SECRET,
         store_tokens=_store_tokens,
-        network_layer=logging_network,
         access_token=tokens.access,
         refresh_token=tokens.refresh,
         refresh_lock=refresh_lock,
     )
-    client = Client(oauth2, network_layer=logging_network)
+    client = Client(oauth2)
     _do_work(client)
 
 
@@ -72,20 +69,6 @@ def _do_work(client):
     # To see token refresh, perhaps put this in a loop (and don't forget to sleep for a bit between requests).
     me = client.user(user_id='me').get()
     items = client.folder('0').get_items(10)
-
-
-class LoggerProxy(Logger):
-    """
-    Proxy for a logger that injects the current PID before log messages.
-    """
-    def __init__(self, logger):
-        self._logger_log = logger._log
-        logger._log = self._log
-        self._preamble = 'PID {0}: '.format(getpid())
-
-    def _log(self, level, msg, args, exc_info=None, extra=None):
-        msg = self._preamble + msg
-        return self._logger_log(level, msg, args, exc_info=exc_info, extra=extra)
 
 
 if __name__ == '__main__':
