+++ docs/source/boxsdk.network.rst
@@ -12,14 +12,6 @@ boxsdk.network.default_network module
     :undoc-members:
     :show-inheritance:
 
-boxsdk.network.logging_network module
--------------------------------------
-
-.. automodule:: boxsdk.network.logging_network
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
 boxsdk.network.network_interface module
 ---------------------------------------
