+++ docs/source/boxsdk.session.rst
@@ -4,6 +4,22 @@ boxsdk.session package
 Submodules
 ----------
 
+boxsdk.session.box_request module
+---------------------------------
+
+.. automodule:: boxsdk.session.box_request
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+boxsdk.session.box_response module
+----------------------------------
+
+.. automodule:: boxsdk.session.box_response
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
 boxsdk.session.box_session module
 ---------------------------------
