+++ test/unit/util/test_log.py
@@ -47,7 +47,7 @@ def test_setup_logging(stream_or_file, debug, expected_log_level, name, mock_log
     with patch('logging.getLogger') as get_logger:
         with patch('logging.open', mock_file_open, create=True):
             get_logger.return_value = mock_logger
-            assert boxsdk.util.log.setup_logging(stream_or_file, debug=debug, name=name) == mock_logger
+            assert boxsdk.util.log.Logging().setup_logging(stream_or_file, debug=debug, name=name) == mock_logger
             get_logger.assert_called_once_with(name)
 
     assert mock_logger.addHandler.call_count == 1
@@ -57,3 +57,22 @@ def test_setup_logging(stream_or_file, debug, expected_log_level, name, mock_log
     if isinstance(stream_or_file, string_types):
         assert mock_file_open.call_count == 1
         assert mock_file_open.call_args[0][:2] == (stream_or_file, 'w')   # Python 3 passes additional args.
+
+
+def test_setup_logging_is_reentrant(mock_logger):
+    mock_file_open = mock_open()
+
+    with patch('logging.getLogger') as get_logger:
+        with patch('logging.open', mock_file_open, create=True):
+            get_logger.return_value = mock_logger
+            assert boxsdk.util.log.setup_logging() == mock_logger
+            get_logger.assert_called_once_with(None)
+            get_logger.return_value = Mock()
+            assert boxsdk.util.log.setup_logging() is None
+
+    assert mock_logger.addHandler.call_count == 1
+    assert isinstance(mock_logger.addHandler.call_args[0][0], logging.Handler)
+    mock_logger.setLevel.assert_called_once()
+
+    if isinstance(stream_or_file, string_types):
+        assert mock_file_open.call_count == 1
